Extract data from a plan

Extract data from a vault connection. The data is first processed with columns translated into more redable variants. The data is then packed into a single tibble with nested data.frames, containing information about:

data_samples for sample metadata

data_samples_community for community (vegetation) data

data_traits for trait data

data_abiotic for abiotic data

Usage

extract_data(
  con,
  return_raw_data = FALSE,
  check_mandatory_references = TRUE,
  verbose = TRUE
)

Arguments

con

A vault connection

return_raw_data

A logical indicating whether to return raw data or without any processing. Default is FALSE.

check_mandatory_references

A logical indicating whether to check for presentce of mandatory reference. Default is TRUE.

verbose

A logical indicating whether output additional information. Default is TRUE.

Value

A data.frame