Select dataset by geographical location

Filter the dataset by geographical location

Usage

select_dataset_by_geo(
  con = NULL,
  long_lim = c(-180, 180),
  lat_lim = c(-90, 90),
  sel_dataset_type = c("vegetation_plot", "fossil_pollen_archive", "gridpoints"),
  verbose = TRUE
)

Arguments

con

A class of vault_pipe

long_lim

A numeric vector of length 2 specifying the longitude limits

lat_lim

A numeric vector of length 2 specifying the latitude limits

sel_dataset_type

(optional) A character vector specifying the dataset types to filter

verbose

A logical value specifying whether to print messages

Value

A class of vault_pipe

Details

There is an option to filter only specific dataset types (as some dataset types may not have lat/long information). To do this, use the sel_dataset_type argument.