Classify taxa

Classify taxa in a SQL table to a specific level based on the classification table in the Vault database.

Usage

classify_taxa(
  data_source = NULL,
  sel_con = NULL,
  to = c("original", "species", "genus", "family")
)

Arguments

data_source

A SQL table that contains taxon_id column.

sel_con

A connection to the Vault database.

to

A character vector that specifies the classification level to classify the taxa. The options are original, species, genus, and family.

Value

A SQL table with the classified taxa.