Pack the data into nested tibbles

Pack the data into nested tibbles for easier access and manipulation. Only specific columns are selected and packed into the following nested tibbles:

data_samples for sample metadata

data_community for community (vegetation) data

data_traits for trait data

data_abiotic for abiotic data When the sel_data does not contain the required columns, the function will skip the respective data type.

Usage

pack_data(sel_data = NULL, verbose = TRUE)

Arguments

sel_data

A tibble with data

verbose

A logical indicating whether to output additional information. Default is TRUE.

Value

A tibble with nested data.frames